Create PhraseOMatic.java
This exercise is a little different to what is suggested on the book, I used a words from the internet to create the name of a dragon randomly from 3 lists. 

Also I imported the java.util.Random class from the java.util package to use nextInt method to show better results.